Fix the problem of SinkIOBuffer getting stuck
